content/html/content/test/test_bug557628.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=557628
-->
<head>
  <title>Test for Bug 557628</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=557628">Mozilla Bug 557628</a>
<p id="display"></p>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 557628 **/

var formAutocompleteTestData = [
  // Default value.
  [ "on" ],
  // Valid values.
  [ "on", "off" ],
  // Invalid values.
  [ "", " ", "foo", "default" ]
];

function checkAttribute(element, name, data)
{
  is(element.getAttribute(name), undefined,
     "By default " + name + " content attribute should be undefined");
  is(element[name], data[0][0],
     "By default " + name + " IDL attribute should be equal to " +
     data[0][0]);

  // Valid values.
  for (i in data[1]) {
    element.setAttribute(name, data[1][i]);
    is(element.getAttribute(name), data[1][i],
       "getAttribute should return the content attribute");
    is(element[name], data[1][i], "When getting, " + name + " IDL attribute " +
       "should be equal to the content attribute if the value is known");
  }

  // Invalid values.
  for (i in data[2]) {
    element.setAttribute(name, data[2][i]);
    is(element.getAttribute(name), data[2][i],
       "getAttribute should return the content attribute");
    is(element[name], data[0][0], "When getting, " + name + " IDL attribute " +
       "should return the default value if the content attribute value isn't known");
  }

  // TODO values.
  for (i in data[3]) {
    element.setAttribute(name, data[3][i]);
    is(element.getAttribute(name), data[3][i],
       "getAttribute should return the content attribute");
    todo_is(element[name], data[3][i], "When getting, " + name + " IDL attribute " +
       "should be equal to the content attribute if the value is known");
  }
}

var form = document.createElement('form');

checkAttribute(form, 'autocomplete', formAutocompleteTestData);

</script>
</pre>
</body>
</html>
